+++ security_auth_center/src/main/java/com/xh/auth/service/User2roleService.java
@@ -15,5 +15,5 @@ import java.util.List;
  */
 public interface User2roleService {
 
-    List<User2role> findByUserId(Integer userId);
+    List<User2role> findByUserId(Long userId);
 }
